tools/cmd/vendor: Use tagged submodule commits
And commit .gitmodules to version control.  For several dependencies,
this the referenced commit is unchanged.  For image-spec and x/net,
we're getting a submodule where we didn't have one before.  For
logrus, the referenced commit is rolling back from v1.0.3-11-g89742ae
to v1.0.3 (the latest tag).  For cli, the referenced commit is rolling
forward from v1.17.0-227-g9e5b048 to v1.20.0 (the latest tag).